new branch creation

+++ benchmark/bubble_sort/scnd.c
@@ -2,8 +2,8 @@
 #include <8051.h>
 #include <values.h>
 
-typedef long TARGET_TYPE;
-typedef long TARGET_INDEX;
+typedef int16_t TARGET_TYPE;
+typedef int16_t TARGET_INDEX;
 
 void swap(TARGET_INDEX index_1, TARGET_INDEX index_2)
 {
